chore:remove old implementation version
