+++ contracts/implementation/OneTimeUsePaymaster.sol
@@ -16,11 +16,14 @@ contract OneTimeUsePaymaster is BasePaymaster, PrepaidGasPool {
     /*///////////////////////////////////////////////////////////////
                               EVENTS
   //////////////////////////////////////////////////////////////*/
-    event UserOpSponsored(
-        bytes32 indexed userOpHash,
+    event UserOpSponsoredWithNullifier(
         address indexed sender,
-        uint256 actualGasCost
+        bytes32 indexed userOpHash,
+        uint256 actualGasCost,
+        uint256 nullifierUsed
     );
+    event RevenueWithdrawn(address withdrawAddress, uint256 amount);
+
     /*///////////////////////////////////////////////////////////////
                               ERRORS
     //////////////////////////////////////////////////////////////*/
@@ -164,7 +167,12 @@ contract OneTimeUsePaymaster is BasePaymaster, PrepaidGasPool {
         // Deduct the JOINING_AMOUNT from the pool's total deposits.
         totalDeposit -= JOINING_AMOUNT;
 
-        emit UserOpSponsored(userOpHash, sender, totalGasCost);
+        emit UserOpSponsoredWithNullifier(
+            sender,
+            userOpHash,
+            totalGasCost,
+            nullifier
+        );
     }
 
     function _pull(
@@ -175,8 +183,25 @@ contract OneTimeUsePaymaster is BasePaymaster, PrepaidGasPool {
         _depositToEntryPoint(msg.value);
     }
 
-    function _push(
-        address _recipient,
-        uint256 _value
-    ) internal virtual override(PrepaidGasPool) {}
+    /// @notice Override to allow revenue withdrawal
+    function _withdrawTo(
+        address payable withdrawAddress,
+        uint256 amount
+    ) internal virtual override(BasePaymaster) {
+        uint256 currentEntryPointDeposit = getDeposit();
+        uint256 revenue = currentEntryPointDeposit - totalDeposit;
+
+        if (amount == 0 || amount > revenue) {
+            revert WithdrawalNotAllowed();
+        }
+
+        entryPoint.withdrawTo(withdrawAddress, amount);
+        emit RevenueWithdrawn(withdrawAddress, amount);
+    }
+
+    /// @notice Get available revenue
+    function getRevenue() public view returns (uint256) {
+        uint256 currentEntryPointDeposit = getDeposit();
+        return currentEntryPointDeposit - totalDeposit;
+    }
 }
